chore: split ClusterSelection component
