+++ src/redux/initialState.ts
@@ -56,8 +56,8 @@ const initialAppConfigState: AppConfig = {
     theme: electronStore.get('appConfig.settings.theme'),
     textSize: electronStore.get('appConfig.settings.textSize'),
     language: electronStore.get('appConfig.settings.language'),
-    loadLastProjectOnStartup: electronStore.get('appConfig.settings.loadLastFolderOnStartup'),
-    isClusterSelectorVisible: electronStore.get('ui.clusterStatusHidden'),
+    loadLastProjectOnStartup: electronStore.get('appConfig.settings.loadLastProjectOnStartup'),
+    isClusterSelectorVisible: electronStore.get('appConfig.settings.isClusterSelectorVisible'),
     hideExcludedFilesInFileExplorer: electronStore.get('appConfig.settings.hideExcludedFilesInFileExplorer'),
     enableHelmWithKustomize: electronStore.get('appConfig.settings.enableHelmWithKustomize'),
   },
